+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 
 ## [2.6.0] - 2020-12-21
 
+### Added
+
+- Print difference between current release and desired release.
+
 ### Changed
 
 - Updated Helm to v3.4.2.